requirements.txt
astroid==2.1.0
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pika==0.12.0
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.2
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
whitenoise==4.1.2
wrapt==1.10.11
